+++ src/plonemeeting/restapi/services/users.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from plone.restapi.services.users.get import UsersGet
-from plonemeeting.restapi.utils import may_access_config_endpoints
+from plonemeeting.restapi.utils import get_poweraccess_configs
 from zope.interface import implementer
 from zope.publisher.interfaces import IPublishTraverse
 
@@ -14,10 +14,10 @@ class PMUsersGet(UsersGet):
         self.query = self.request.form.copy()
 
     def has_permission_to_query(self):
-        return may_access_config_endpoints()
+        return get_poweraccess_configs()
 
     def has_permission_to_enumerate(self):
-        return may_access_config_endpoints()
+        return get_poweraccess_configs()
 
     def has_permission_to_access_user_info(self):
-        return may_access_config_endpoints()
+        return get_poweraccess_configs()